+++ packages/storybook/stories/ElementsStories/Button.stories.tsx
@@ -1,13 +1,17 @@
+import { useState } from "react";
+
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { Button } from "@sikka/hawa/elements/button";
 import { Logos } from "@sikka/hawa/elements/logos";
 import { Tooltip } from "@sikka/hawa/elements/tooltip";
 
+import { setLocale } from "../../translations/i18n";
+
 const meta = {
   title: "Elements/Button",
   component: Button,
-  parameters: { layout: "centered" }
+  parameters: { layout: "centered" },
 } satisfies Meta<typeof Button>;
 
 export default meta;
@@ -57,7 +61,7 @@ export const Variations: Story = {
         </div>
       </div>
     );
-  }
+  },
 };
 
 export const Sizes: Story = {
@@ -118,5 +122,57 @@ export const Sizes: Story = {
         </div>
       </div>
     );
-  }
+  },
+};
+
+export const Playground: Story = {
+  render: (args: any, globals: any) => {
+    const locale = globals.globals?.locale === "ar" ? "ar" : "en";
+    const direction = locale === "ar" ? "rtl" : "ltr";
+    setLocale(locale);
+    const [isLoading, setIsLoading] = useState(false);
+    const [sent, setSent] = useState(false);
+    const mockSubmit = async (data: any) => {
+      return new Promise((resolve, reject) => {
+        setIsLoading(true);
+        setTimeout(() => {
+          resolve("Submission successful");
+          setIsLoading(false);
+          setSent(true);
+        }, 2000);
+      });
+    };
+    return (
+      <div className="hawa-flex hawa-flex-col hawa-gap-2">
+        <h1>Variations</h1>
+        <div className="hawa-flex hawa-flex-row hawa-flex-wrap hawa-gap-2">
+          <Button variant="default">Default</Button>
+          <Button variant="outline" isLoading={isLoading}>
+            Outline
+          </Button>
+          <Button variant="destructive">Destructive</Button>
+          <Button variant="ghost">Ghost</Button>
+          <form
+            className="hawa-flex hawa-flex-col hawa-gap-4"
+            id="test"
+            onSubmit={(e) => {
+              e.preventDefault();
+              mockSubmit("test");
+            }}
+          >
+            <Button
+              variant="light"
+              type="submit"
+              form="test"
+              isLoading={isLoading}
+            >
+              Inside Form
+            </Button>
+          </form>{" "}
+          <Button variant="secondary">Secondary</Button>
+          <Button variant="link">Link</Button>
+        </div>
+      </div>
+    );
+  },
 };